Rust changes

+++ rs/backend/src/accounts_store/histogram.rs
@@ -97,5 +97,5 @@ impl Add<&Account> for AccountsStoreHistogram {
 
 /// Determines which log base 2 bucket a count falls into.
 fn log2_bucket(count: usize) -> u32 {
-    (1 << usize::ilog2(count * 2 + 1)) - 1
+    ((1u64 << usize::ilog2(count * 2 + 1)) - 1) as u32
 }
